move accelerator legacy tests (#5948)
