+++ dockers/tpu-tests/tpu_test_cases.jsonnet
@@ -23,7 +23,7 @@ local tputests = base.BaseTest {
       cd pytorch-lightning
       coverage run --source=pytorch_lightning -m pytest -v --capture=no \
           pytorch_lightning/utilities/xla_device_utils.py \
-          tests/accelerators/legacy/test_tpu_backend.py \
+          tests/accelerators/test_tpu_backend.py \
           tests/models/test_tpu.py
       test_exit_code=$?
       echo "\n||| END PYTEST LOGS |||\n"
