+++ tests/accelerators/__init__.py
@@ -1,0 +1,12 @@
+try:
+    from dtrun.launcher import DDPLauncher
+except ImportError:
+
+    class DDPLauncher:
+
+        def run(cmd_line, **kwargs):
+
+            def inner(func):
+                pass
+
+            return inner
