+++ tests/accelerators/test_ddp_spawn.py
