+++ tests/trainer/properties/test_get_model.py
@@ -17,7 +17,7 @@ import pytest
 import torch
 
 from pytorch_lightning import Trainer
-from tests.accelerators.legacy import DDPLauncher
+from tests.accelerators import DDPLauncher
 from tests.helpers.boring_model import BoringModel
